Улучшает нейминг модулей, изменяет вид экспорта
